created separate function for updatePOI validator
