+++ server/routes/api/trailRoutes.js
@@ -6,7 +6,10 @@ const { fieldsUpload } = require("../../config/imageUpload");
 const { listTrails, saveTrail } = require("../../controllers/trailControllers");
 const { addPOI, updatePOI } = require("../../controllers/poiControllers");
 const { saveTrailValidator } = require("../../validators/saveTrailValidator");
-const { addPOIValidator } = require("../../validators/pOIValidators");
+const {
+  addPOIValidator,
+  updatePOIValidator,
+} = require("../../validators/pOIValidators");
 
 // base url: https://gamma.lessthangeeky.com/api/trails/ +
 
@@ -29,7 +32,14 @@ router.post(
   addPOI,
   listTrails
 );
-router.post("/updatePOI", verifyToken, fieldsUpload, updatePOI);
+router.post(
+  "/updatePOI",
+  verifyToken,
+  fieldsUpload,
+  updatePOIValidator,
+  updatePOI,
+  listTrails
+);
 
 // get "/:trailId", <-- returns all trail details (POIs, Hazards, etc.)
 // post "/:trailId/close", verifyToken,
