+++ server/validators/pOIValidators.js
@@ -1,6 +1,7 @@
 const { body } = require("express-validator");
 
 const { VALID_IMAGE_TYPES } = require("../config/imageUpload");
+const { PointsOfInterest } = require("../models");
 
 exports.addPOIValidator = [
   // Points of Interest validation
@@ -47,4 +48,54 @@ exports.addPOIValidator = [
     .optional()
     .isFloat()
     .toFloat(),
+];
+
+exports.updatePOIValidator = [
+  // Points of Interest validation
+  body().custom((value, { req }) => {
+    // Add the files back in to the req.body so that they can be treated normally in validation
+    const files = req.files.image;
+    if (files) {
+      value.files = files[0];
+    }
+    return true;
+  }),
+  // required fields
+  body("pointsOfInterestId")
+    .exists()
+    .isInt()
+    .toInt()
+    .bail()
+    .custom(async (value) => {
+      const poi = await PointsOfInterest.findByPk(value);
+      console.log("***** Check if POI exists");
+      if (!poi) {
+        console.log("****** POI doesn't exists");
+        throw new Error("pointsOfInterestId doesn't exist");
+      }
+      console.log("****** POI exists");
+      return true;
+    }),
+  // optional fields
+  body("description", "Invalid data type, must be a string")
+    .optional()
+    .isString()
+    .trim()
+    .escape(),
+  body("files")
+    .optional()
+    .custom((value) => {
+      // check valid mime types
+      const mimetypeArr = value.mimetype.split("/");
+
+      return (
+        mimetypeArr[0] === "image" &&
+        VALID_IMAGE_TYPES.indexOf(mimetypeArr[1]) > -1
+      );
+    })
+    .withMessage(`Files must be of type: .${VALID_IMAGE_TYPES.join(", .")}`),
+  body("isActive", "Point of Interest isActive must be true/false")
+    .optional()
+    .isBoolean()
+    .toBoolean(),
 ];
